externalise wind calculation and processing from client class
